requirements.txt
# Web Scraping & HTTP
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium==4.18.1

# Database
sqlalchemy==2.0.27

# Web Framework
streamlit>=1.24.0

# CRM Integration
hubspot-api-client>=7.7.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# NLP
nltk>=3.8.1

# Email Validation
email-validator>=2.0.0

# Additional dependencies
scrapy>=2.11.0
openai>=1.0.0
sendgrid>=6.10.0
python-dateutil>=2.8.2
urllib3>=2.0.0
certifi>=2023.7.22
charset-normalizer>=3.2.0
idna>=3.4
lxml>=4.9.3
soupsieve>=2.4.1
twisted>=22.10.0
w3lib>=2.1.2
zope.interface>=6.0 